% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survAUC-package.R
\docType{package}
\name{survAUC-package}
\alias{survAUC-package}
\alias{survAUC}
\title{survAUC: Estimators of Prediction Accuracy for Time-to-Event Data}
\description{
Provides a variety of functions to estimate time-dependent true/false positive rates and AUC curves from a set of censored survival data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/survAUC/}
  \item \url{https://github.com/fbertran/survAUC/}
  \item Report bugs at \url{https://github.com/fbertran/survAUC/issues/}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Sergej Potapov
  \item Werner Adler
  \item Matthias Schmid
}

}
\keyword{internal}
