% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survAUC-package.R
\docType{package}
\name{survAUC-package}
\alias{survAUC}
\alias{survAUC-package}
\title{survAUC: Estimators of prediction accuracy for time-to-event data.}
\description{
The package provides a variety of functions to estimate time-dependent true/false positive rates and AUC curves from a set of censored survival data.
}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me}

Authors:
\itemize{
  \item Sergej Potapov \email{sergej.potapov@gmail.com}
  \item Werner Adler
  \item Matthias Schmid
}

}
\keyword{internal}
